DEV-8 Add supervisor_service testing (#2)
* Add failing supervisor_service test

This test returns an error regarding `Chef::Resource::SupervisorService`.

* Make the supervisor_service resource compatible with Chef 13-14

- Moved the provider code to the resource file
- Reworked the enable and disable action
- Bumped the cookbook version

* Fix some cookstyle offenses

* Add more testing to the supervisor_service resource

Although the tests were passing, I wasn't convinced that the resource was
working as expected which led to some reworking of the custom resource. While
there are still some holes in the test coverage, I believe it should be good
enough to get things working again.